feat: V1.0 full-stack A-stock quantitative trading platform
- Backend: FastAPI + SQLite WAL, akshare data pipeline, 40-factor alpha library
- Engine: Volume-price scoring (60%) + character scoring (40%) → composite signals
- Backtest: V1 fixed stop-loss/trailing-stop, V2 ATR dynamic stop + volatility sizing
- Frontend: React 19 + TypeScript + lightweight-charts, dark theme
- Pages: Dashboard, Ranking (year/month/IPO tabs), Backtest (dual strategy), Alerts
- Risk: DynamicStopManager, VolatilityPositionSizer, MarketRegimeFilter
- Strategy: MeanReversionStrategy, LightGBM model, WalkForwardValidator
- Scripts: historical scan, daily update, extended data fetch, pipeline runners

+++ backend/engine/backtest/metrics.py
@@ -59,7 +59,7 @@ def profit_loss_ratio(trades: list[dict]) -> float:
     losses = [abs(t["pnl"]) for t in trades if t["pnl"] < 0]
     if not losses or not wins:
         return 0.0
-    return round(np.mean(wins) / np.mean(losses), 4)
+    return round(float(np.mean(wins) / np.mean(losses)), 4)
 
 
 def compute_all_metrics(equity_curve: pd.Series, trades: list[dict]) -> dict:
